+++ src/ModelStates/IsStateManager.php
@@ -4,33 +4,36 @@ namespace MLL\LaravelUtils\ModelStates;
 
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\MorphTo;
-use Illuminate\Support\Collection as SupportCollection;
+use Illuminate\Support\Collection;
 
 /**
  * @property-read HasStateManagerInterface&Model $stateable
  *
  * Virtual
- * @property-read SupportCollection<State> $canTransitionTo
+ * @property-read Collection<array-key, State> $canTransitionTo
  */
 trait IsStateManager
 {
-    /** @return MorphTo<HasStateManagerInterface&Model, self> */
+    /** @return MorphTo<HasStateManagerInterface&Model, $this> */
     abstract public function stateable(): MorphTo;
 
     abstract public static function stateColumnName(): string;
 
-    /** @return SupportCollection<class-string<State>, State> */
-    public function getCanTransitionToAttribute(): SupportCollection
+    /** @return Collection<class-string<State>, State> */
+    public function getCanTransitionToAttribute(): Collection
     {
         $stateable = $this->stateable;
         $stateable->refresh();
 
+        $state = $stateable->state;
+        assert($state instanceof State);
+
         return $stateable->stateClass()::config()
-            ->possibleNextStates($stateable->state);
+            ->possibleNextStates($state);
     }
 
-    /** @return SupportCollection<array-key, Transition> */
-    public function possibleTransitions(): SupportCollection
+    /** @return Collection<array-key, Transition> */
+    public function possibleTransitions(): Collection
     {
         $stateable = $this->stateable;